Add support for pre 3.11 f-strings
It works by maintaining a preferred quote order for the depth of recursion.
Generally, single is preferred as the outer quote with double quotes as the inner.
If there's a >2 depth of f-strings, triple quotes are preferred on the outer with singles inside.

This allows for the maximum f-string nesting depth allowed in Py3.11 (4 levels)